Organizado o projeto: Padrao_de_aqruitetura_MVVM_separando_a_parte_logica_da_minha_view_principal
