+++ app/src/main/java/com/example/androidapi/network/configuration/ui/MainActivity.kt
@@ -2,11 +2,14 @@ package com.example.androidapi
 
 import android.os.Bundle
 import androidx.activity.enableEdgeToEdge
+import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
+import androidx.lifecycle.Observer
 import com.example.androidapi.network.configuration.EventsInterface
 import com.example.androidapi.network.configuration.RetrofitConfiguration
+import com.example.androidapi.network.configuration.ViewModelActiviy.FilmesViewModel
 import com.example.androidapi.network.configuration.model.FilmesResponse
 import retrofit2.Call
 import retrofit2.Callback
@@ -14,6 +17,8 @@ import retrofit2.Response
 
 class MainActivity : AppCompatActivity() {
     val urlBasic = "http://www.omdbapi.com"
+    private val filmesViewModel: FilmesViewModel by viewModels()
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         enableEdgeToEdge()
@@ -23,24 +28,15 @@ class MainActivity : AppCompatActivity() {
             v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom)
             insets
         }
-        getFilmes()
 
+        filmesViewModel.filmes.observe(this, Observer { filmes ->
 
-    }
-    fun getFilmes() {
-        val retrofitClient = RetrofitConfiguration.getRetrofitIntance(urlBasic)
-        val interfaceGet = retrofitClient.create(EventsInterface::class.java)
-        val callBack = interfaceGet.getFilmes()
-
-        callBack.enqueue(object: Callback<FilmesResponse> {
-            override fun onFailure(call: Call<FilmesResponse>, t: Throwable) {
-                val error = t.message
-
-            }
-            override fun onResponse(call: Call<FilmesResponse>, response: Response<FilmesResponse>) {
-                response.body()
-
-            }
         })
+
+        filmesViewModel.getFilmes(urlBasic)
+
     }
+
+
+
 }
