+++ app/src/main/res/layout/activity_main.xml
@@ -6,7 +6,7 @@
         android:id="@+id/main"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".MainActivity">
+        tools:context=".network.configuration.ui.MainActivity">
     <TextView
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
